Order shards by all counts
This is the first stab at ranking a shards on their popularity. The
algorithm is simply forks + stars + subscribers + watchers. It
certainly works if two shards have significantly different counts. But
until we have a lot of repos that have very similar counts we won't
really get to see how representative this method is.