Make popular_first method safer
This accomplishes two things:

* Include the table along with the column when ordering by the sum of
multiple columns
* Ensure that the columns actually exist (according to the Query object)

This is the only way I could find to get the table name from an
existing object. Shard::TABLE_NAME said it was an undefined constant.

I added the table to everything but the first column because the
order_by method is expecting one column adds the table name to the
front of the query passed in. If I added the table name it would print
it twice: `shards.shards.stargazers_count `

Each column is actually a LuckyRecord::Criteria which has the column
name stored.

This way, if the table name changes or one of the columns changes or is
removed, crystal won't let this compile.