Add Categories model/form/migration
A Category simply has a title. I chose "title" instead of "name"
because the difference between the two can be very ambiguous. I tend to
reserve "name" for if the thing I'm representing has a personality.
Think Person, User, Pet, AutonomousCompanion. Everything else should
have a title. The exception is if there's some preexisting model you
need to adhere to that uses name.