Use type inference to refactor ShardCreationForm
The Shards::ShardCreationForm module includes a `render` method that
takes a single ShardForm as its only argument. Because this module is
included in the shard new/index pages, calling render with a ShardForm
will render the whole thing. This is much nicer than wrapping the render
method to just pass along the single parmeter to `render_shard_form`