Make a shard's description nillable
We grab the description from the GitHub repo, and it might not be set.

The migration simply removes the NULL constraint on the database. The
reverse migration sets the default to be an empty string, then re-adds
the NULL constraint.

I needed to change the type for description from `String` to `String?`
and finally, ensure that if there is no description the ShowPage
doesn't crash.