Add .local to .gitignore
I like to have a space where I can put temporary files or personal files
relating to a project, but that don't need to be committed. Screenshots,
notes, etc.

I use a .local directory for this, but I'm all ears for other names or
ways to accomplish the same thing.